+++ lib/pay.rb
@@ -9,6 +9,7 @@ module Pay
   autoload :NanoId, "pay/nano_id"
   autoload :Payment, "pay/payment"
   autoload :Receipts, "pay/receipts"
+  autoload :Currency, "pay/currency"
 
   # Payment processors
   autoload :Braintree, "pay/braintree"